fix: uint32_t for lastSleepMinutes (was uint16_t, overflowed at 3-month hibernate)
Winter hibernate returns 129600 min which exceeds uint16_t max (65535), wrapping
to 64064 in stored/logged values. The actual sleep duration was unaffected since
it is computed from nextWake - now in seconds, but the field was misleading.

Also fix ARCHITECTURE.md wave sampling duration (160s / 1600 samples, not 3 min / 1800).

+++ src/battery.cpp
@@ -58,7 +58,7 @@ bool handleUndervoltageProtection() {
     uint32_t now = (uint32_t)time(NULL);
     uint32_t candidate = (now >= 24 * 3600 ? now : 0) + (uint32_t)sleepMinutes * 60UL;
     uint32_t nextWake = adjustNextWakeUtcForQuietHours(candidate);
-    rtcState.lastSleepMinutes = (uint16_t)sleepMinutes;
+    rtcState.lastSleepMinutes = (uint32_t)sleepMinutes;
     rtcState.lastNextWakeUtc = nextWake;
     // Print local next wake for convenience
     if (rtcState.lastNextWakeUtc >= 24 * 3600) {
